requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.31.0
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
pyyaml>=6.0.0 